Widen context traversal depth to usize
collect_root_rules carried BFS depth as u16 and incremented depth + 1
unguarded, so a syntactically valid corpus with a parent chain deeper
than 65535 panicked under the release overflow checks this crate
enables. AdrId admits far more IDs than u16 can index as a depth, so
the type could not represent every valid traversal.

Widen the traversal depth and EmittedRule.depth to usize rather than
introducing a typed depth cap: the cap would change the call chain's
result type, while the widening is confined to context.rs plus the
EmittedRule field, which AFM-0026:R2 makes crate-private and
non-breaking to reshape. depth is never read by the renderer.

Findings adr-fmt-nwi#5.

+++ src/output.rs
@@ -68,7 +68,7 @@ pub struct EmittedRule {
     pub rule_id: String,
     pub text: String,
     pub layer: u8,
-    pub depth: u16,
+    pub depth: usize,
 }
 
 /// Render a `--refs` report as a compact markdown bullet list.
